+++ src/command/globalConfig.ts
@@ -1,8 +1,9 @@
 import { window, workspace } from 'vscode';
-import { isTokenValid } from '../lib/valid';
 
 import { Job } from './command';
+import { isTokenValid } from '../lib/valid';
 import { getExecuteRobotByInput } from './projectConfig';
+import { DEFAULT_TEMPLATES, TemplateType } from '../lib/projectTemplate';
 
 export async function setApiToken(): Promise<Job> {
     let token = await window.showInputBox({
@@ -17,7 +18,9 @@ export async function setApiToken(): Promise<Job> {
         return Job.undone;
     }
 
-    workspace.getConfiguration("hamibot-assistant").update("ApiToken", token, true);
+    await workspace
+        .getConfiguration("hamibot-assistant")
+        .update("ApiToken", token, true);
     return Job.done;
 }
 
@@ -30,7 +33,9 @@ export async function setShowOfflineRobot(): Promise<Job> {
         return Job.undone;
     }
 
-    workspace.getConfiguration("hamibot-assistant").update("OfflineRobot", choose === "是", true);
+    await workspace
+        .getConfiguration("hamibot-assistant")
+        .update("OfflineRobot", choose === "是", true);
     return Job.done;
 }
 
@@ -41,11 +46,43 @@ export async function setDefaultExecuteRobot(): Promise<Job> {
         return Job.undone;
     }
 
-    workspace.getConfiguration("hamibot-assistant.defaultExecuteRobot")
+    await workspace
+        .getConfiguration("hamibot-assistant.defaultExecuteRobot")
         .update("robotInfo", {
             name: robot.name,
             id: robot._id
         }, true);
 
+    return Job.done;
+}
+
+export async function choseDefaultTemplate(): Promise<Job> {
+    let choose = await window.showQuickPick(DEFAULT_TEMPLATES.map((value) => {
+        let mark;
+        switch (value.config.type) {
+            case TemplateType.local:
+                mark = "💻 [本地] ";
+                break;
+            case TemplateType.remote:
+                mark = "🌏 [远程] ";
+                break;
+            default:
+                mark = "";
+        }
+
+        return {
+            label: mark + value.name,
+            detail: value.description,
+            config: value.config
+        };
+    }));
+
+    if (!choose) {
+        return Job.undone;
+    }
+
+    await workspace
+        .getConfiguration("hamibot-assistant")
+        .update("projectTemplate", choose.config, true);
     return Job.done;
 }